+++ docs/CHANGELOG.md
@@ -4,3 +4,4 @@ Changelog
 0.0.1 - Unreleased
 -------------------------
 - Init: Initialize Module "Text editor"
+- Enh #1: Store file version on update content